Move tests base settings to baseapp_core tests dir
