+++ .github/workflows/project-workflow.yml
@@ -41,7 +41,6 @@ jobs:
       - name: black
         id: black
         if: steps.isort.outcome == 'success'
-        # TODO: Check if can check black only on the package
         run: black --check ./${{ inputs.project }}
       # Test
       - run: echo "🧪 Checking ${{ inputs.project }} Tests"
@@ -49,12 +48,11 @@ jobs:
       - name: Install tests requirements
         if: steps.black.outcome == 'success'
         id: requirements_installation
-        # TODO: Check if it is baseapp-core to avoid installing twice
         run: |
           python -m pip install --upgrade pip
           pip install -r baseapp-core/testproject/requirements.txt
           requirements_path="${{ inputs.project }}/testproject/requirements.txt"
-          if [ -f ${requirements_path} ]; then pip install -r ${requirements_path}; else echo 'No requirements.txt file found in tests directory'; fi
+          if [ -f ${requirements_path} ]; then pip install -r ${requirements_path}; else echo 'No requirements.txt file found in testproject directory'; fi
       - name: pytest
         if: steps.requirements_installation.outcome == 'success'
         run: |
