+++ baseapp-auth/testproject/settings.py
@@ -1,26 +1,13 @@
 from collections import OrderedDict
-from pathlib import Path
 
-from baseapp_core.settings.base import *  # noqa
 from baseapp_core.settings.env import env
+from baseapp_core.tests.settings import *  # noqa
 from celery.schedules import crontab
 
-# Build paths inside the project like this: BASE_DIR / 'subdir'.
-BASE_DIR = Path(__file__).resolve().parent.parent
 APPS_DIR = BASE_DIR
 
-
-# Quick-start development settings - unsuitable for production
-# See https://docs.djangoproject.com/en/3.2/howto/deployment/checklist/
-
-# SECURITY WARNING: keep the secret key used in production secret!
-SECRET_KEY = "secret-for-test-app"
-
-# SECURITY WARNING: don't run with debug turned on in production!
-DEBUG = True
-
-ALLOWED_HOSTS = []
-
+if DEFAULT_AUTO_FIELD:
+    del DEFAULT_AUTO_FIELD
 
 # Application definition
